Moved components / pages into separate directories again
