--- client/src/components/auth/Login.js
@@ -1,110 +1,0 @@
-import React, { Component } from "react";
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
-import { loginUser } from "../../actions/authActions";
-
-class Login extends Component {
-  constructor() {
-    super();
-    this.state = {
-      email: "",
-      password: "",
-      errors: {}
-    };
-  }
-
-  componentDidMount() {
-    // If logged in and user navigates to Login page, should redirect them to dashboard
-    if (this.props.auth.isAuthenticated) {
-      this.props.history.push("/dashboard");
-    }
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.auth.isAuthenticated) {
-      this.props.history.push("/dashboard");
-    }
-
-    if (nextProps.errors) {
-      this.setState({
-        errors: nextProps.errors
-      });
-    }
-  }
-
-  onChange = e => {
-    this.setState({ [e.target.id]: e.target.value });
-  };
-
-  onSubmit = e => {
-    e.preventDefault();
-
-    const userData = {
-      email: this.state.email,
-      password: this.state.password
-    };
-
-    this.props.loginUser(userData);
-  };
-
-  render() {
-    const { errors } = this.state;
-
-    return (
-      <section>
-        <h4>Login below</h4>
-        <form noValidate onSubmit={this.onSubmit}>
-          <div>
-            <input
-              onChange={this.onChange}
-              value={this.state.email}
-              error={errors.email}
-              id="email"
-              type="email"
-            />
-            <label htmlFor="email">Email</label>
-            <span>
-              {errors.email}
-              {errors.emailnotfound}
-            </span>
-          </div>
-          <div>
-            <input
-              onChange={this.onChange}
-              value={this.state.password}
-              error={errors.password}
-              id="password"
-              type="password"
-            />
-            <label htmlFor="password">Password</label>
-            <span>
-              {errors.password}
-              {errors.passwordincorrect}
-            </span>
-          </div>
-          <span>{errors.status}</span>
-          <div>
-            <button type="submit">Login</button>
-          </div>
-        </form>
-      </section>
-    );
-  }
-}
-
-Login.propTypes = {
-  loginUser: PropTypes.func.isRequired,
-  auth: PropTypes.object.isRequired,
-  errors: PropTypes.object.isRequired
-};
-
-const mapStateToProps = state => ({
-  auth: state.auth,
-  errors: state.errors,
-  isLoading: state.auth.loading
-});
-
-export default connect(
-  mapStateToProps,
-  { loginUser }
-)(Login);
