--- client/src/components/layout/Landing.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-
-const Landing = () => (
-  <section>
-    <h2>Landing</h2>
-    <p><Link to="/login">login</Link></p>
-  </section>
-);
-
-export default Landing;
